// src/reviews/dto/reviews.dto.ts
interface DtoActionsReview {
  typeWork: string;
  gradeTotal: number;
  gradeQuality: number;
  gradeMaterials: number;
  gradePrice: number;
  gradeExperience: number;
  gradeDeadlines: number;
  gradeCommunication: number;
  descDignity: string;
  descFlaws: string;
  descReview: string;
  images: string;
}

export type { DtoActionsReview };